// angular/todo/v1/src/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** DateTime scalar */
  DateTime: any;
  /** JSON scalar */
  Json: any;
  /** Any scalar. DO NOT USE except in extremely rare cases where it is unavoidable to do so */
  Any: any;
  /** Long type */
  Long: any;
};
// ...
export type Query = {
  __typename?: 'Query';
  /** Find objects of the given type which match the query. Use a conditional fragment to select fields as in https://graphql.org/learn/schema/#union-types */
  find: HypiFilterConnection;
  /** Gets a single object by its ID (hypi.id) */
  get?: Maybe<HypiRootAggregate>;
  aggregate?: Maybe<HypiAggregationType>;
  login?: Maybe<AccessToken>;
  loginByEmail?: Maybe<AccessToken>;
  me?: Maybe<PermissionDescription>;
  hasPermission?: Maybe<Array<Maybe<Scalars['Boolean']>>>;
};
// ...
export type NotificationCtx = {
  __typename?: 'NotificationCtx';
  hypi?: Maybe<Hypi>;
  /** The type that the notification applies to */
  type?: Maybe<Scalars['String']>;
  targetAccount?: Maybe<Scalars['ID']>;
};
// ...
export type RequestTemplate = {
  __typename?: 'RequestTemplate';
  hypi?: Maybe<Hypi>;
  name: Scalars['String'];
  request?: Maybe<Scalars['String']>;
  response?: Maybe<Scalars['String']>;
};
// ...
export type Todos = {
  __typename?: 'Todos';
  hypi?: Maybe<Hypi>;
  id?: Maybe<Scalars['Int']>;
  title?: Maybe<Scalars['String']>;
  description?: Maybe<Scalars['String']>;
  completed?: Maybe<Scalars['Boolean']>;
};
// ...
export type HypiRootAggregate = Todos | PageInfo | HypiResultEdge | HypiFilterConnection | HypiEnv | Pair | AggInt | AggFloat | AggOtherScalar | Script | RequestTemplate | NotificationCtx | Notification | Url | Currency | Coordinate | GeoEnvelope | Language | Address | PersonName | Phone | Email | Password | RemoteLogin | LoginAttempt | BruteForceDetectionOptions | OAuth2AuthorizedClient | AuthClient | AbacPolicy | AbacTag | Image | EmailVerification | EmailTemplate | EmailSendingAttempt | PasswordReminder | Webhook | WebhookResponse | LogMessage | GraphQlRef | WorkflowStepData | WorkflowStep | AccessToken | StorageCounter | PermissionDescription | Hypi | Country | Account | Person | Organisation | OAuthProvider | Realm | Group | Role | RolePolicy | ClientPolicy | TimePolicy | AggregatedPolicy | GroupPolicy | AccountPolicy | RealmPolicy | RealmLink | Permission | File | Video | EmailMessage | Workflow | WorkflowSession | Counter | Gauge | ServerlessResponse;

/** An object available as "env" in all scripts */
export type HypiEnv = {
  __typename?: 'HypiEnv';
  hypi?: Maybe<Hypi>;
  apiHost?: Maybe<Scalars['String']>;
  websocketHost?: Maybe<Scalars['String']>;
};
// ...
export type Notification = {
  __typename?: 'Notification';
  hypi?: Maybe<Hypi>;
  message?: Maybe<Scalars['String']>;
  ctx?: Maybe<NotificationCtx>;
};
// ...
export type HypiResultEdge = {
  __typename?: 'HypiResultEdge';
  hypi?: Maybe<Hypi>;
  node: HypiRootAggregate;
  cursor: Scalars['ID'];
};
// ...
export type PageInfo = {
  __typename?: 'PageInfo';
  hypi?: Maybe<Hypi>;
  hasPreviousPage: Scalars['Boolean'];
  hasNextPage: Scalars['Boolean'];
  startCursor: Scalars['ID'];
  /** relay modern - https://relay.dev/graphql/connections.htm#note-95f8a */
  endCursor: Scalars['ID'];
  /**
   * relay modern
   * If present, is the page size that is used to generate the previousOffsets and nextOffsets if they're present
   */
  pageLimit?: Maybe<Scalars['Int']>;
  /** If present, contains the pagination cursors that can be used to get the previous N pages */
  previousOffsets?: Maybe<Array<Scalars['String']>>;
  /** If present, contains the pagination cursors that can be used to get the next N pages */
  nextOffsets?: Maybe<Array<Scalars['String']>>;
};
// ...
export type AggFloat = {
  __typename?: 'AggFloat';
  hypi?: Maybe<Hypi>;
  /** The value of the aggregated field for each group */
  groupValues?: Maybe<Array<Maybe<Pair>>>;
  avg?: Maybe<Scalars['Float']>;
  count?: Maybe<Scalars['Int']>;
  max?: Maybe<Scalars['Float']>;
  min?: Maybe<Scalars['Float']>;
  sum?: Maybe<Scalars['Float']>;
};
// ...
export type Mutation = {
  __typename?: 'Mutation';
  /** Removes the link between two objects WITHOUT deleting either object */
  unlink: Scalars['Boolean'];
  /** Creates a relationship between two existing objects via some field on the source type */
  link: Scalars['Boolean'];
  /** Inserts or updates a set of values. To update an existing value provide the hypi.id */
  upsert: Array<Hypi>;
  /** Marks any object that matches the filter as NOT trash if they were previous trashed with the 'trash' function */
  untrash: Scalars['Int'];
  math: Array<Hypi>;
  /** Deletes any objects matching the query up to a maximum of 25 (regardless of the first/last arguments if they are bigger than 25) */
  delete: Scalars['Int'];
  /** Deletes the given scalar values from the object with the given ID from the field in the specified type */
  deleteScalars: Scalars['Int'];
  /** Marks any object that matches the filter as trash. Those objects will not be returned in get or find queries unless they're marked as not trash later. Maximum of 25 will be trashed (regardless of the first/last arguments if they are bigger than 25) */
  trash: Scalars['Int'];
  /** Used to perform registration of a new realm. Normally only Hypi directly calls this */
  createRealm?: Maybe<Hypi>;
  /** Creates a new account which can be used to login and perform various actions in a Realm */
  createAccount?: Maybe<Hypi>;
};
// ...
export type Phone = {
  __typename?: 'Phone';
  hypi?: Maybe<Hypi>;
  number: Scalars['String'];
  country?: Maybe<Country>;
  code?: Maybe<Scalars['String']>;
};
// ...
export type GeoEnvelope = {
  __typename?: 'GeoEnvelope';
  hypi?: Maybe<Hypi>;
  p1: Coordinate;
  p2: Coordinate;
};
// ...
export type Currency = {
  __typename?: 'Currency';
  hypi?: Maybe<Hypi>;
  name: Scalars['String'];
  code: Scalars['String'];
  symbol: Scalars['String'];
};
// ...
export type HypiFilterConnection = {
  __typename?: 'HypiFilterConnection';
  hypi?: Maybe<Hypi>;
  edges?: Maybe<Array<HypiResultEdge>>;
  pageInfo: PageInfo;
};
// ...
export type PermissionDescription = {
  __typename?: 'PermissionDescription';
  hypi?: Maybe<Hypi>;
  roles?: Maybe<Array<Role>>;
  groups?: Maybe<Array<Group>>;
  organisations?: Maybe<Array<Organisation>>;
  realms?: Maybe<Array<Realm>>;
  permissions?: Maybe<Array<Maybe<Permission>>>;
};
// ...
export type AggOtherScalar = {
  __typename?: 'AggOtherScalar';
  hypi?: Maybe<Hypi>;
  /** The value of the aggregated field for each group */
  groupValues?: Maybe<Array<Maybe<Pair>>>;
  count?: Maybe<Scalars['Int']>;
};
// ...
export type Pair = {
  __typename?: 'Pair';
  hypi?: Maybe<Hypi>;
  key?: Maybe<Scalars['String']>;
  value?: Maybe<Scalars['String']>;
};
// ...
export type Coordinate = {
  __typename?: 'Coordinate';
  hypi?: Maybe<Hypi>;
  x: Scalars['Float'];
  y: Scalars['Float'];
};
// ...
export type AggInt = {
  __typename?: 'AggInt';
  hypi?: Maybe<Hypi>;
  /** The value of the aggregated field for each group */
  groupValues?: Maybe<Array<Maybe<Pair>>>;
  avg?: Maybe<Scalars['Float']>;
  count?: Maybe<Scalars['Int']>;
  max?: Maybe<Scalars['Int']>;
  min?: Maybe<Scalars['Int']>;
  sum?: Maybe<Scalars['Int']>;
};
// ...
export type AccessToken = {
  __typename?: 'AccessToken';
  hypi?: Maybe<Hypi>;
  sessionToken?: Maybe<Scalars['String']>;
  sessionExpires?: Maybe<Scalars['Long']>;
  errorCode?: Maybe<Scalars['String']>;
  errorMsg?: Maybe<Scalars['String']>;
};
// ...
export type Script = {
  __typename?: 'Script';
  hypi?: Maybe<Hypi>;
  type: TanType;
  name: Scalars['String'];
  body: Scalars['String'];
};
// ...
export type DeleteMutationVariables = Exact<{
  arcql: Scalars['String'];
  clearArrayReferences?: Maybe<Scalars['Boolean']>;
}>;


export type DeleteMutation = (
  { __typename?: 'Mutation' }
  & Pick<Mutation, 'delete'>
);
// ...
export type TodosFieldsFragment = (
  { __typename?: 'Todos' }
  & Pick<Todos, 'title' | 'description'>
  & { hypi?: Maybe<(
    { __typename?: 'Hypi' }
    & Pick<Hypi, 'id'>
  )> }
);

export type UpsertMutationVariables = Exact<{
  values: HypiUpsertInputUnion;
}>;


export type UpsertMutation = (
  { __typename?: 'Mutation' }
  & { upsert: Array<(
    { __typename?: 'Hypi' }
    & Pick<Hypi, 'id'>
  )> }
);

export type TodosQueryVariables = Exact<{
  arcql: Scalars['String'];
}>;


export type TodosQuery = (
  { __typename?: 'Query' }
  & { find: (
    { __typename?: 'HypiFilterConnection' }
    & { edges?: Maybe<Array<(
      { __typename?: 'HypiResultEdge' }
      & { node: (
        { __typename?: 'Todos' }
        & TodosFieldsFragment
      ) | { __typename?: 'PageInfo' } | { __typename?: 'HypiResultEdge' } | { __typename?: 'HypiFilterConnection' } | { __typename?: 'HypiEnv' } | { __typename?: 'Pair' } | { __typename?: 'AggInt' } | { __typename?: 'AggFloat' } | { __typename?: 'AggOtherScalar' } | { __typename?: 'Script' } | { __typename?: 'RequestTemplate' } | { __typename?: 'NotificationCtx' } | { __typename?: 'Notification' } | { __typename?: 'URL' } | { __typename?: 'Currency' } | { __typename?: 'Coordinate' } | { __typename?: 'GeoEnvelope' } | { __typename?: 'Language' } | { __typename?: 'Address' } | { __typename?: 'PersonName' } | { __typename?: 'Phone' } | { __typename?: 'Email' } | { __typename?: 'Password' } | { __typename?: 'RemoteLogin' } | { __typename?: 'LoginAttempt' } | { __typename?: 'BruteForceDetectionOptions' } | { __typename?: 'OAuth2AuthorizedClient' } | { __typename?: 'AuthClient' } | { __typename?: 'ABACPolicy' } | { __typename?: 'ABACTag' } | { __typename?: 'Image' } | { __typename?: 'EmailVerification' } | { __typename?: 'EmailTemplate' } | { __typename?: 'EmailSendingAttempt' } | { __typename?: 'PasswordReminder' } | { __typename?: 'Webhook' } | { __typename?: 'WebhookResponse' } | { __typename?: 'LogMessage' } | { __typename?: 'GraphQLRef' } | { __typename?: 'WorkflowStepData' } | { __typename?: 'WorkflowStep' } | { __typename?: 'AccessToken' } | { __typename?: 'StorageCounter' } | { __typename?: 'PermissionDescription' } | { __typename?: 'Hypi' } | { __typename?: 'Country' } | { __typename?: 'Account' } | { __typename?: 'Person' } | { __typename?: 'Organisation' } | { __typename?: 'OAuthProvider' } | { __typename?: 'Realm' } | { __typename?: 'Group' } | { __typename?: 'Role' } | { __typename?: 'RolePolicy' } | { __typename?: 'ClientPolicy' } | { __typename?: 'TimePolicy' } | { __typename?: 'AggregatedPolicy' } | { __typename?: 'GroupPolicy' } | { __typename?: 'AccountPolicy' } | { __typename?: 'RealmPolicy' } | { __typename?: 'RealmLink' } | { __typename?: 'Permission' } | { __typename?: 'File' } | { __typename?: 'Video' } | { __typename?: 'EmailMessage' } | { __typename?: 'Workflow' } | { __typename?: 'WorkflowSession' } | { __typename?: 'Counter' } | { __typename?: 'Gauge' } | { __typename?: 'ServerlessResponse' } }
    )>> }
  ) }
);
// ...
export const DeleteDocument = gql`
    mutation delete($arcql: String!, $clearArrayReferences: Boolean = false) {
  delete(type: Todos, arcql: $arcql, clearArrayReferences: $clearArrayReferences)
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class DeleteMutationService extends Apollo.Mutation<DeleteMutation, DeleteMutationVariables> {
    document = DeleteDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
// ...
export const UpsertDocument = gql`
    mutation upsert($values: HypiUpsertInputUnion!) {
  upsert(values: $values) {
    id
  }
}
    `;

  @Injectable({
    providedIn: 'root'
  })
  export class UpsertMutationService extends Apollo.Mutation<UpsertMutation, UpsertMutationVariables> {
    document = UpsertDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }
export const TodosDocument = gql`
    query todos($arcql: String!) {
  find(type: Todos, arcql: $arcql) {
    edges {
      node {
        ...TodosFields
      }
    }
  }
}
    ${TodosFieldsFragmentDoc}`;

  @Injectable({
    providedIn: 'root'
  })
  export class TodosQueryService extends Apollo.Query<TodosQuery, TodosQueryVariables> {
    document = TodosDocument;
    
    constructor(apollo: Apollo.Apollo) {
      super(apollo);
    }
  }